requirements.txt
streamlit==1.32.2
openai==1.14.2
python-dotenv==1.0.1
pandas==2.2.1
